[ELY-1425] Re-distribute the JASPIC implementation classes allowing for a public / private API split.
